requirements.txt
pyinstaller>=5.13.0
Pillow>=9.0.0

